Organized the InOut folder
