+++ Echo.Core/src/InOut/Images/Serializer.cs
@@ -3,7 +3,7 @@ using System.IO;
 using Echo.Core.Textures.Colors;
 using Echo.Core.Textures.Grids;
 
-namespace Echo.Core.Textures.Serialization;
+namespace Echo.Core.InOut.Images;
 
 /// <summary>
 /// A serialization interface for saving and loading <see cref="TextureGrid{T}"/> into and out of <see cref="Stream"/>.
