+++ Echo.Terminal/src/EchoTI.cs
@@ -1,6 +1,7 @@
 ﻿using Echo.Core.Common.Compute;
 using Echo.Core.Evaluation.Operation;
 using Echo.Core.InOut;
+using Echo.Core.InOut.Scenes;
 using Echo.Core.Scenic.Hierarchies;
 using Echo.Core.Scenic.Preparation;
 using Echo.Terminal.Application;
